:recycle: refactor : 공통 컴포넌트와 유틸 및 API 설정 구조 정리
