+++ src/shared/components/InfoNotice.jsx
