+++ src/shared/components/WarningNotice.jsx
