+++ src/shared/styles/InfoNotice.css
