+++ src/shared/util/constant.js
